Shuffle some APIs around
