+++ crates/api/src/trampoline/global.rs
@@ -1,5 +1,5 @@
 use super::create_handle::create_handle;
-use crate::runtime::StoreInstanceHandle;
+use crate::trampoline::StoreInstanceHandle;
 use crate::Store;
 use crate::{GlobalType, Mutability, Val};
 use anyhow::{bail, Result};
