+++ crates/api/src/trampoline/memory.rs
@@ -1,6 +1,6 @@
 use super::create_handle::create_handle;
 use crate::externals::{LinearMemory, MemoryCreator};
-use crate::runtime::StoreInstanceHandle;
+use crate::trampoline::StoreInstanceHandle;
 use crate::Store;
 use crate::{Limits, MemoryType};
 use anyhow::Result;
